requirements/base.txt
Django>=2.0,<3.0
django-environ==0.4.5
psycopg2==2.7.4 --no-binary psycopg2
redis>=3.2.0
django-redis==4.10.0
celery==4.2.1
flower==0.9.2
tornado>=4.2.0,<6.0.0

# Django REST Framework
djangorestframework==3.9.0
django-filter==2.0.0

#Documentation
django-rest-swagger
